refactor: created a new file for colorspace conversion and moved the functions from utils.js into it
