+++ src/core/change-brightness.js
@@ -1,4 +1,4 @@
-import utils from "../utils.js";
+import {rgbToHsl, hslToRgb} from "../core/colorspace-conversion.js";
 
 /**
  * Function to change the brightness of an image
@@ -10,9 +10,9 @@ function changeBrightness(pixelArray, factor) {
     if (factor == 0)
         return pixelArray;
     for (let i = 0; i < pixelArray.length; i += 4) {
-        const hsl = utils.rgbToHsl(pixelArray[i], pixelArray[i + 1], pixelArray[i + 2]);
+        const hsl = rgbToHsl(pixelArray[i], pixelArray[i + 1], pixelArray[i + 2]);
         hsl.l = Math.max(0, Math.min(100, hsl.l += factor));
-        const rgb = utils.hslToRgb(hsl.h, hsl.s, hsl.l);
+        const rgb = hslToRgb(hsl.h, hsl.s, hsl.l);
 
         pixelArray[i] = rgb.r;
         pixelArray[i + 1] = rgb.g;
